requirements.txt
flask
flask-cors
supabase
python-dotenv
bcrypt
pyjwt
requests
